feat: rename service to taskmanager
