--- scripts/services/start-redis.sh
@@ -1,3 +1,0 @@
-#! /bin/bash
-
-go run ./cmd/redis/main.go
