+++ scripts/services/start-taskmanager.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+
+go run ./cmd/taskmanager/main.go
